UBER-695: Allow front service configuration
