+++ server/front/src/__start.ts
@@ -14,134 +14,5 @@
 // limitations under the License.
 //
 
-import { MinioService } from '@hcengineering/minio'
-import { setMetadata } from '@hcengineering/platform'
-import serverToken from '@hcengineering/server-token'
-import { start } from '.'
-
-const defaultLanguage = process.env.DEFAULT_LANGUAGE ?? 'en'
-const SERVER_PORT = parseInt(process.env.SERVER_PORT ?? '8080')
-
-const transactorEndpoint = process.env.TRANSACTOR_URL
-if (transactorEndpoint === undefined) {
-  console.error('please provide transactor url')
-  process.exit(1)
-}
-
-const elasticUrl = process.env.ELASTIC_URL
-if (elasticUrl === undefined) {
-  console.error('please provide elastic url')
-  process.exit(1)
-}
-
-const minioEndpoint = process.env.MINIO_ENDPOINT
-if (minioEndpoint === undefined) {
-  console.error('please provide minio endpoint')
-  process.exit(1)
-}
-
-const minioAccessKey = process.env.MINIO_ACCESS_KEY
-if (minioAccessKey === undefined) {
-  console.error('please provide minio access key')
-  process.exit(1)
-}
-
-const minioSecretKey = process.env.MINIO_SECRET_KEY
-if (minioSecretKey === undefined) {
-  console.error('please provide minio secret key')
-  process.exit(1)
-}
-
-const minio = new MinioService({
-  endPoint: minioEndpoint,
-  port: 9000,
-  useSSL: false,
-  accessKey: minioAccessKey,
-  secretKey: minioSecretKey
-})
-
-const accountsUrl = process.env.ACCOUNTS_URL
-if (accountsUrl === undefined) {
-  console.error('please provide accounts url')
-  process.exit(1)
-}
-
-const uploadUrl = process.env.UPLOAD_URL
-if (uploadUrl === undefined) {
-  console.error('please provide upload url')
-  process.exit(1)
-}
-
-const collaboratorUrl = process.env.COLLABORATOR_URL
-if (collaboratorUrl === undefined) {
-  console.error('please provide collaborator url')
-  process.exit(1)
-}
-
-const gmailUrl = process.env.GMAIL_URL
-if (gmailUrl === undefined) {
-  console.error('please provide gmail url')
-  process.exit(1)
-}
-
-const calendarUrl = process.env.CALENDAR_URL
-if (calendarUrl === undefined) {
-  console.error('please provide calendar service url')
-  process.exit(1)
-}
-
-const telegramUrl = process.env.TELEGRAM_URL
-if (telegramUrl === undefined) {
-  console.error('please provide telegram url')
-  process.exit(1)
-}
-
-const rekoniUrl = process.env.REKONI_URL
-if (rekoniUrl === undefined) {
-  console.error('please provide rekoni url')
-  process.exit(1)
-}
-
-const modelVersion = process.env.MODEL_VERSION
-if (modelVersion === undefined) {
-  console.error('please provide model version requirement')
-  process.exit(1)
-}
-
-const serverSecret = process.env.SERVER_SECRET
-if (serverSecret === undefined) {
-  console.log('Please provide server secret')
-  process.exit(1)
-}
-
-const title = process.env.TITLE
-
-setMetadata(serverToken.metadata.Secret, serverSecret)
-
-const config = {
-  transactorEndpoint,
-  elasticUrl,
-  minio,
-  accountsUrl,
-  uploadUrl,
-  modelVersion,
-  collaboratorUrl,
-  gmailUrl,
-  telegramUrl,
-  rekoniUrl,
-  calendarUrl,
-  title,
-  defaultLanguage
-}
-console.log('Starting Front service with', config)
-const shutdown = start(config, SERVER_PORT)
-
-const close = (): void => {
-  console.trace('Exiting from server')
-  console.log('Shutdown request accepted')
-  shutdown()
-  process.exit(0)
-}
-
-process.on('SIGINT', close)
-process.on('SIGTERM', close)
+import { startFront } from './starter'
+startFront()
